backend/requirements_improved.txt
# Requirements améliorés pour l'application ESA

# Core Flask
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23
Werkzeug==3.0.1

# Security
python-dotenv==1.0.0
email-validator==2.1.0
cryptography==41.0.7

# PDF & Documents
reportlab==4.0.7
Pillow==10.1.0
openpyxl==3.1.2  # Pour Excel

# QR Code
qrcode==7.4.2

# Date & Time
python-dateutil==2.8.2

# Caching & Performance
redis==5.0.1
celery==5.3.4  # Pour tâches asynchrones

# Email
sendgrid==6.11.0  # Alternative à Flask-Mail

# SMS
twilio==8.10.0  # Pour notifications SMS

# API Documentation
flask-restx==1.3.0  # Pour Swagger/OpenAPI

# Monitoring & Logging
sentry-sdk[flask]==1.38.0  # Pour error tracking
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
faker==20.1.0  # Pour données de test

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# File Upload
flask-uploads==0.2.1

# Background Tasks
APScheduler==3.10.4  # Pour tâches planifiées

# Webhooks
flask-webhook==0.1.0

# Compression
flask-compress==1.14

# Rate Limiting (alternative)
slowapi==0.1.9


